moved time and snowflake functions into a module
